[Fix] WebTransportCompany - Soft-delete validation

+++ src/microservices/thirdParties/v1/controllers/web/transportCompanies.js
@@ -202,6 +202,16 @@ exports.postDelete = async (req, res, next) => {
         // // ! Pendiente: Validar permisos del usuario
         // createdBy: createdBy.id,
       },
+      // ! Es necesario borrar primero las rutas?
+      include: [
+        {
+          model: db.TransportRoute,
+          attributes: ["id"],
+          required: false,
+        },
+      ],
+      attributes: ["id"],
+      paranoid: true,
     });
 
     if (companyInDb === null) {
@@ -210,6 +220,12 @@ exports.postDelete = async (req, res, next) => {
         message: `Transport company does not found`,
       };
     }
+
+    if (companyInDb.TransportRoutes != 0)
+      throw {
+        status: StatusCodes.UNPROCESSABLE_ENTITY,
+        message: `Transport company has related routes`,
+      };
 
     await companyInDb.destroy();
